move helper function to other file
